ANDKR-1966 | 업장 혜택 UI 추가

+++ dailyHOTEL/app/src/main/java/com/daily/dailyhotel/entity/StayDetailk.kt
@@ -73,7 +73,7 @@ class StayDetailk : Configurations() {
 
     class TrueReviewInformation {
         var ratingCount: Int = 0
-        var ratingPercent: Float = 0.0f
+        var ratingPercent: Int = 0
         var showRating: Boolean = false
 
         var review: PrimaryReview? = null
